+++ src/tokens_select_mt.cpp
@@ -20,8 +20,8 @@ Text keep(Text tokens,
     if(tokens.size() == 0) return {}; // return empty vector for empty text
     
     Text tokens_copy(tokens.size(), 0);
-    for(int span = span_max; span >= 1; span--){ // substitution starts from the longest sequences
-        for(int i = 0; i < tokens.size() - (span - 1); i++){
+    for (size_t span = span_max; span >= 1; span--){ // substitution starts from the longest sequences
+        for (size_t i = 0; i < tokens.size() - (span - 1); i++){
             Ngram ngram(tokens.begin() + i, tokens.begin() + i + span);
             bool is_in = set_words.find(ngram) != set_words.end();
             if(is_in){
@@ -43,8 +43,8 @@ Text remove(Text tokens,
     Text tokens_copy(tokens.size(), 0);
     unsigned int filler = std::numeric_limits<unsigned int>::max(); // use upper limit as a filler
     bool match = false;
-    for(int span = span_max; span >= 1; span--){ // substitution starts from the longest sequences
-        for(int i = 0; i < tokens.size() - (span - 1); i++){
+    for (size_t span = span_max; span >= 1; span--){ // substitution starts from the longest sequences
+        for (size_t i = 0; i < tokens.size() - (span - 1); i++){
             Ngram ngram(tokens.begin() + i, tokens.begin() + i + span);
             bool is_in = set_words.find(ngram) != set_words.end();
             if(is_in){
@@ -75,15 +75,15 @@ struct select_mt : public Worker{
     void operator()(std::size_t begin, std::size_t end){
         //Rcout << "Range " << begin << " " << end << "\n";
         if(mode == 1){
-            for (int h = begin; h < end; h++){
+            for (size_t h = begin; h < end; h++){
                 output[h] = keep(input[h], span_max, set_words, padding);
             }
         }else if(mode == 2){
-            for (int h = begin; h < end; h++){
+            for (size_t h = begin; h < end; h++){
                 output[h] = remove(input[h], span_max, set_words, padding);
             }
         }else{
-            for (int h = begin; h < end; h++){
+            for (size_t h = begin; h < end; h++){
                 output[h] = input[h];
             }
         }
@@ -117,7 +117,7 @@ List qatd_cpp_tokens_select(List texts_,
 
     SetNgrams set_words;
     int span_max = 0;
-    for(int g = 0; g < words.size(); g++){
+    for (size_t g = 0; g < words.size(); g++){
         if(has_na(words[g])) continue;
         Ngram word = words[g];
         set_words.insert(word);
